# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pixel WEB SERVER",
    "version": "1.0.0",
    "description": "Документація до беку"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "https",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
    "/admin/users": {
      "get": {
        "summary": "Отримати інформацію про користувачів",
        "description": "Цей ендпоінт дозволяє адміністраторам отримати інформацію про користувачів. Підтримує запити за ID користувача, ID замовлення або фільтрацію на основі параметрів запиту.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "ID користувача, якого потрібно отримати. Якщо вказано, повертає вказаного користувача.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "objectId"
            }
          },
          {
            "name": "orderId",
            "in": "query",
            "description": "ID замовлення для отримання інформації про конкретного користувача. Вимагає ID користувача.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "objectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успішно отримано інформацію про користувача або замовлення",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "fullName": {
                            "type": "string"
                          },
                          "phone": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "city": {
                            "type": "string"
                          },
                          "history": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "format": "objectId"
                            }
                          }
                        }
                      }
                    },
                    "order": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string"
                          },
                          "phone": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Користувача або замовлення не знайдено",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "error": "Користувача не знайдено"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "error": "Внутрішня помилка сервера"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Видалення користувачів",
        "description": "Цей ендпоінт дозволяє видаляти користувачів з бази даних. Можна видалити всіх користувачів або лише вибраних, вказавши їх ID.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "description": "ID користувачів для видалення. Якщо не вказано, видаляються всі користувачі.",
            "schema": {
              "type": "string",
              "example": "603c6b2f7f2e1b001f64792c"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Користувачі успішно видалені",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Повідомлення про успішне видалення.",
                      "example": "Deleted"
                    }
                  },
                  "required": [
                    "message"
                  ]
                },
                "example": {
                  "message": "Deleted"
                }
              }
            }
          },
          "400": {
            "description": "Помилка запиту через некоректні дані",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Опис помилки",
                      "example": "Invalid request data"
                    }
                  },
                  "required": [
                    "error"
                  ]
                },
                "example": {
                  "error": "Invalid request data"
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Опис помилки",
                      "example": "Server error"
                    }
                  },
                  "required": [
                    "error"
                  ]
                },
                "example": {
                  "error": "Server error"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/export": {
      "get": {
        "summary": "Експорт даних у CSV",
        "description": "Цей ендпоінт дозволяє адміністраторам експортувати дані з вказаної колекції у формат CSV. Потрібно вказати параметр `collection`. За бажанням, можна також вказати параметр `id` для експорту конкретних документів.",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "ID документів для експорту. Якщо не вказано, експортуються всі документи з колекції.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "objectId"
            }
          },
          {
            "name": "collection",
            "in": "query",
            "description": "Назва колекції, з якої потрібно експортувати дані.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успішний експорт даних у форматі CSV",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Помилка через відсутність назви колекції",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Collection name is required"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/users/orders": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Видалення замовлень користувача",
        "description": "Цей ендпоінт дозволяє користувачам видаляти свої замовлення. Можна видалити всі замовлення або вибрані за ідентифікаторами.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Список ідентифікаторів замовлень, які потрібно видалити, розділених комами. Якщо не вказано, будуть видалені всі замовлення користувача.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "60d21b4667d0d8992e610c85,60d21b4667d0d8992e610c86"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Замовлення успішно видалено",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "All Orders deleted"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/calls": {
      "put": {
        "tags": [
          "Anon"
        ],
        "summary": "Залишити номер телефону",
        "description": "Цей запит дозволяє додати новий запис про дзвінок до бази даних адміністратора. Для цього потрібно надати номер телефону та текст повідомлення.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  }
                },
                "required": [
                  "phone"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Запис про дзвінок успішно додано",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "message": "We will call you"
                }
              }
            }
          },
          "400": {
            "description": "Неправильний номер телефону",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Send correct phone number"
                }
              }
            }
          },
          "500": {
            "description": "Помилка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      }
    }
  }
}